maintain the login

+++ application/controllers/Sarpras.php
@@ -13,32 +13,51 @@ class Sarpras extends CI_Controller {
 
     public function index()
 	{
+if ($this->session->userdata('logged_in')) {
         $data =[
             "data" => $this->M_home->getAllData("sarpras"),
             "main_view" => "v_sarpras"
         ];
-		$this->load->view('v_layout',$data);
+    $this->load->view('v_layout',$data);
+    }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     }
 
     public function tambah()
 	{
+if ($this->session->userdata('logged_in')) {
         $data =[
             "main_view" => "v_tambah_sarpras"
         ];
-		$this->load->view('v_layout',$data);
+    $this->load->view('v_layout',$data);
+    }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
   }
 
     public function ubah($id)
 	{
+    if ($this->session->userdata('logged_in')) {
         $data =[
             "data" => $this->M_home->getData(array("idSarpras" => $id), "sarpras")->row(),
             "main_view" => "v_ubah_sarpras"
         ];
-		$this->load->view('v_layout',$data);
+    $this->load->view('v_layout',$data);
+    }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     }
 
     public function add()
     {
+if ($this->session->userdata('logged_in')) {
     //   $this->form_validation->set_rules('', '', '');
     //   if ($this->form_validation->run() == TRUE) {
                 
@@ -55,13 +74,19 @@ class Sarpras extends CI_Controller {
     //           $this->session->set_flashdata('notif', 'One of required input is empty');
     //           redirect(''); /* need to modified */
     //   }
+    }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     }
 
     public function edit($id)
     {
+
     //   $this->form_validation->set_rules('', '', '');
     //   if ($this->form_validation->run() == TRUE) {
-                
+                if ($this->session->userdata('logged_in')) {
                     if ($this->M_sarpras->edit($id)) {
                       $this->session->set_flashdata('type', 'success');
                       $this->session->set_flashdata('notif', 'Sukses ubah data');
@@ -71,6 +96,11 @@ class Sarpras extends CI_Controller {
                       $this->session->set_flashdata('notif', 'Gagal ubah data');
                       redirect('sarpras'); /* need to modified */
                     }
+                    }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     //   } else {
     //           $this->session->set_flashdata('notif', 'One of required input is empty');
     //           redirect(''); /* need to modified */
@@ -79,6 +109,7 @@ class Sarpras extends CI_Controller {
 
     public function delete($id)
     {
+if ($this->session->userdata('logged_in')) {
       if($this->M_sarpras->delete($id)){
         $this->session->set_flashdata('type', 'success');
         $this->session->set_flashdata('notif', 'Sukses hapus data');
@@ -87,6 +118,11 @@ class Sarpras extends CI_Controller {
       }
       
       redirect('sarpras');
+      }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     }
 }
