+++ application/controllers/TahapanPelaksanaan.php
@@ -12,12 +12,19 @@ class TahapanPelaksanaan extends CI_Controller {
     }
 
     public function index(){
+if ($this->session->userdata('logged_in')) {
         $tipe_ = "pemeriksaanbersama";
         redirect('/tahapanpelaksanaan/data/'.$tipe_);
+        }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     }
 
     public function data($tipe_)
 	{
+        if ($this->session->userdata('logged_in')) {
         if($tipe_ == "pemeriksaanbersama"){
             $tipe = "Pemeriksaan Bersama";
         } else if($tipe_ == "pengajuanpersyaratan"){
@@ -35,7 +42,12 @@ class TahapanPelaksanaan extends CI_Controller {
             "data" => $this->M_home->getAllDataTahapan($tipe,"tahapanPelaksanaan"),
             "main_view" => "v_tahapanPelaksanaan"
         ];
-		$this->load->view('v_layout',$data);
+        $this->load->view('v_layout',$data);
+        }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     }
 
  
@@ -43,6 +55,7 @@ class TahapanPelaksanaan extends CI_Controller {
 
 
     public function tambah($tipe_){
+    if ($this->session->userdata('logged_in')) {
         $formatFile = "";
         $tipe = "";
         
@@ -69,10 +82,16 @@ class TahapanPelaksanaan extends CI_Controller {
             "formatFile" => $formatFile, // untuk format file
             "main_view" => "v_tambah_tahapanPelaksanaan"
         ];
-		$this->load->view('v_layout',$data);       
+        $this->load->view('v_layout',$data); 
+           }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }   
     }
 
     public function ubah($tipe_,$id){
+    if ($this->session->userdata('logged_in')) {
         $formatFile = "";
         $tipe = "";
         
@@ -100,13 +119,18 @@ class TahapanPelaksanaan extends CI_Controller {
             "data" => $this->M_home->getData(array("idPelaksanaan" => $id),"tahapanPelaksanaan")->row(),
             "main_view" => "v_ubah_tahapanPelaksanaan"
         ];
-		$this->load->view('v_layout',$data);       
+        $this->load->view('v_layout',$data);  
+          }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }   
     }
     public function add($tipe_)
     {
     //   $this->form_validation->set_rules('', '', '');
     //   if ($this->form_validation->run() == TRUE) {
-                
+                if ($this->session->userdata('logged_in')) {
                     $initialize = $this->upload->initialize(array(
                         'upload_path' => './assets/file/',
                         'allowed_types' => 'jpg|jpeg|png|pdf'
@@ -122,7 +146,11 @@ class TahapanPelaksanaan extends CI_Controller {
                             $this->session->set_flashdata('notif', 'Gagal tambah data');
                             redirect('tahapanPelaksanaan/data/'.$tipe_); /* need to modified */
                             }
-                   
+                   }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     //   } else {
     //           $this->session->set_flashdata('notif', 'One of required input is empty');
     //           redirect(''); /* need to modified */
@@ -133,7 +161,7 @@ class TahapanPelaksanaan extends CI_Controller {
     {
     //   $this->form_validation->set_rules('', '', '');
     //   if ($this->form_validation->run() == TRUE) {
-                
+                if ($this->session->userdata('logged_in')) {
                     $initialize = $this->upload->initialize(array(
                         'upload_path' => './assets/file/',
                         'allowed_types' => 'jpg|jpeg|png|pdf'
@@ -149,7 +177,11 @@ class TahapanPelaksanaan extends CI_Controller {
                             $this->session->set_flashdata('notif', 'Gagal ubah data');
                             redirect('tahapanPelaksanaan/data/'.$tipe_); /* need to modified */
                             }
-                    
+                    }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     //   } else {
     //           $this->session->set_flashdata('notif', 'One of required input is empty');
     //           redirect(''); /* need to modified */
@@ -158,6 +190,7 @@ class TahapanPelaksanaan extends CI_Controller {
 
     public function delete($tipe_, $id)
     {
+if ($this->session->userdata('logged_in')) {
       if($this->M_tahapanPelaksanaan->delete($id)){
         $this->session->set_flashdata('type', 'success');
         $this->session->set_flashdata('notif', 'Sukses hapus data');
@@ -166,9 +199,15 @@ class TahapanPelaksanaan extends CI_Controller {
       }
       
       redirect('tahapanPelaksanaan/data/'.$tipe_); /* need to modified */
+      }else{
+            $this->session->set_flashdata('notif', 'Anda belum login');
+            redirect('auth');
+            
+          }
     }
 
     public function getTahapanPelaksanaan($tipe_){
+if ($this->session->userdata('logged_in')) {
         if($tipe_ == "pemeriksaanbersama"){
             $tipe = "Pemeriksaan Bersama";
         } else if($tipe_ == "pengajuanpersyaratan"){
@@ -181,6 +220,10 @@ class TahapanPelaksanaan extends CI_Controller {
             $tipe = "Kontrak Kritis";
         }
         echo json_encode($this->M_home->getAllDataTahapan($tipe,"tahapanPelaksanaan"));
+        }else{
+            echo "anda tidak punya akses.";
+            
+          }
     }
 }
